refactor: change package(deprecated) - TaskDetails, TaskType, Crawling
