--- src/main/java/com/cnu/simple/work/TaskType.java
@@ -1,5 +1,0 @@
-package com.cnu.simple.work;
-
-public enum TaskType {
-    CRAWLING
-}
