+++ src/main/java/com/cnu/simple/work/deprecated/Crawling.java
@@ -1,0 +1,11 @@
+package com.cnu.simple.work.deprecated;
+
+import jakarta.persistence.*;
+import lombok.Getter;
+
+@Entity
+@Getter
+@DiscriminatorValue("CRAWLING")
+public class Crawling extends TaskDetails {
+
+}
