+++ src/main/java/com/cnu/simple/work/deprecated/TaskType.java
@@ -1,0 +1,5 @@
+package com.cnu.simple.work.deprecated;
+
+public enum TaskType {
+    CRAWLING
+}
